converted examples thru end of Chapter 11
Also, went back and stuck a line into earlier examples to create a non-functioning root view controller in Empty Application template projects where there's no interface and where the user is expected to look at the Console to pick up an NSLog message. This is because otherwise an error message appears complaining about the lack of a root view controller, which just gets in the way.