Improved page view controller example.
Removed the unnecessary complication of the custom container view. The page view is the root view in this example, just as one would do with a tab bar controller or a navigation controller; the pages themselves determine the rotation rules, again just as with the contained views of a navigation controller.